backend/requirements.txt
Django==4.2.24
djangorestframework==3.16.1
django-cors-headers==4.8.0
psycopg2-binary==2.9.9
